+++ electron/main/ipc/ipc-file.ts
@@ -1,64 +1,31 @@
-import type { SongMetadata } from "@native/tools";
-import { app, BrowserWindow, dialog, ipcMain, shell } from "electron";
-import FastGlob from "fast-glob";
-import type { Options as GlobOptions } from "fast-glob/out/settings";
-import { parseFile } from "music-metadata";
-import { access, mkdir, readdir, readFile, stat, unlink, writeFile } from "node:fs/promises";
-import { basename, dirname, extname, isAbsolute, join, relative, resolve } from "node:path";
-import pLimit from "p-limit";
-import type { MusicTrack } from "../database/LocalMusicDB";
-import { ipcLog } from "../logger";
+import { BrowserWindow, ipcMain } from "electron";
+import { join } from "node:path";
 import { LocalMusicService } from "../services/LocalMusicService";
-import { useStore } from "../store";
-import { getFileID, getFileMD5, metaDataLyricsArrayToLrc } from "../utils/helper";
-import { loadNativeModule } from "../utils/native-loader";
-
-type toolModule = typeof import("@native/tools");
-const tools: toolModule = loadNativeModule("tools.node", "tools");
+import { MusicFileService } from "../services/MusicFileService";
+import { DownloadService } from "../services/DownloadService";
+import { FileService } from "../services/FileService";
+import type { MusicTrack } from "../database/LocalMusicDB";
 
 /**
  * 文件相关 IPC
  */
 const initFileIpc = (): void => {
-  /** 本地音乐服务 */
   const localMusicService = new LocalMusicService();
-
-  // Store active download tasks: ID -> DownloadTask instance
-  const activeDownloads = new Map<number, any>();
-
-  /**
-   * 获取全局搜索配置
-   * @param cwd 当前工作目录
-   */
-  const globOpt = (cwd?: string): GlobOptions => ({
-    cwd,
-    caseSensitiveMatch: false,
-  });
+  const musicFileService = new MusicFileService();
+  const downloadService = new DownloadService();
+  const fileService = new FileService();
 
   // 检查文件是否存在
   ipcMain.handle("file-exists", async (_, path: string) => {
-    try {
-      await access(path);
-      return true;
-    } catch {
-      return false;
-    }
+    return fileService.fileExists(path);
   });
 
   // 保存文件
   ipcMain.handle(
     "save-file",
     async (_, args: { path: string; content: string; encoding?: BufferEncoding }) => {
-      try {
-        const { path, content, encoding } = args;
-        const dir = dirname(path);
-        await mkdir(dir, { recursive: true });
-        await writeFile(path, content, { encoding: encoding || "utf-8" });
-        return { success: true };
-      } catch (err) {
-        ipcLog.error("Failed to save file:", err);
-        throw err;
-      }
+      const success = await fileService.saveFile(args.path, args.content, args.encoding);
+      return { success };
     },
   );
 
@@ -66,7 +33,7 @@ const initFileIpc = (): void => {
   ipcMain.handle(
     "get-default-dir",
     (_event, type: "documents" | "downloads" | "pictures" | "music" | "videos"): string => {
-      return app.getPath(type);
+      return fileService.getDefaultDir(type);
     },
   );
 
@@ -74,9 +41,7 @@ const initFileIpc = (): void => {
   ipcMain.handle("local-music-sync", async (event, dirs: string[]) => {
     try {
       // 获取封面目录路径
-      const store = useStore();
-      const localCachePath = join(store.get("cachePath"), "local-data");
-      const coverDir = join(localCachePath, "covers");
+      const { coverDir } = localMusicService.paths;
 
       const processTracksCover = (tracks: MusicTrack[]) => {
         return tracks.map((track) => {
@@ -113,7 +78,7 @@ const initFileIpc = (): void => {
     } catch (err) {
       const errorMessage = err instanceof Error ? err.message : "Unknown error";
       // 如果正在扫描中
-      if (errorMessage === "SCAN_IN_PROGRESS") {
+      if (errorMessage === LocalMusicService.ERROR_SCAN_IN_PROGRESS) {
         return { success: false, message: "扫描正在进行中，请稍候" };
       }
       // 错误信号
@@ -124,685 +89,89 @@ const initFileIpc = (): void => {
 
   // 遍历音乐文件
   ipcMain.handle("get-music-files", async (_, dirPath: string) => {
-    try {
-      // 校验路径有效性
-      if (!dirPath || dirPath.trim() === "") {
-        ipcLog.warn("⚠️ Empty directory path provided, skipping");
-        return [];
-      }
-      // 规范化路径
-      const filePath = resolve(dirPath).replace(/\\/g, "/");
-      // 检查目录是否存在
-      try {
-        await access(filePath);
-      } catch {
-        ipcLog.warn(`⚠️ Directory not accessible: ${filePath}`);
-        return [];
-      }
-      console.info(`📂 Fetching music files from: ${filePath}`);
-      // 音乐文件扩展名
-      const musicExtensions = [
-        "mp3",
-        "wav",
-        "flac",
-        "aac",
-        "webm",
-        "m4a",
-        "ogg",
-        "aiff",
-        "aif",
-        "aifc",
-      ];
-      // 查找指定目录下的所有音乐文件
-      const musicFiles = await FastGlob(`**/*.{${musicExtensions.join(",")}}`, globOpt(filePath));
-
-      // 限制并发数
-      const limit = pLimit(10);
-
-      // 解析元信息（使用 allSettled 防止单个文件失败影响整体）
-      const metadataPromises = musicFiles.map((file) =>
-        limit(async () => {
-          const fullPath = join(dirPath, file);
-          try {
-            // 处理元信息 (跳过封面解析以提升速度)
-            const { common, format } = await parseFile(fullPath, { skipCovers: true });
-            // 获取文件状态信息（大小和创建时间）
-            const fileStat = await stat(fullPath);
-            const ext = extname(fullPath);
-
-            return {
-              id: getFileID(fullPath),
-              name: common.title || basename(fullPath, ext),
-              artists: common.artists?.[0] || common.artist,
-              album: common.album || "",
-              alia: common.comment?.[0]?.text || "",
-              duration: (format?.duration ?? 0) * 1000,
-              size: (fileStat.size / (1024 * 1024)).toFixed(2),
-              path: fullPath,
-              quality: format.bitrate ?? 0,
-              // 文件创建时间（用于排序）
-              createTime: fileStat.birthtime.getTime(),
-              replayGain: {
-                trackGain: common.replaygain_track_gain?.ratio,
-                trackPeak: common.replaygain_track_peak?.ratio,
-                albumGain: common.replaygain_album_gain?.ratio,
-                albumPeak: common.replaygain_album_peak?.ratio,
-              },
-            };
-          } catch (err) {
-            ipcLog.warn(`⚠️ Failed to parse file: ${fullPath}`, err);
-            return null;
-          }
-        }),
-      );
-      const metadataResults = await Promise.all(metadataPromises);
-      // 过滤掉解析失败的文件，并按创建时间降序排序（最新的在前）
-      return metadataResults
-        .filter((item): item is NonNullable<typeof item> => item !== null)
-        .sort((a, b) => b.createTime - a.createTime);
-    } catch (error) {
-      ipcLog.error("❌ Error fetching music metadata:", error);
-      return [];
-    }
+    return musicFileService.getMusicFiles(dirPath);
   });
 
   // 获取音乐元信息
   ipcMain.handle("get-music-metadata", async (_, path: string) => {
-    try {
-      const filePath = resolve(path).replace(/\\/g, "/");
-      const { common, format } = await parseFile(filePath);
-      return {
-        // 文件名称
-        fileName: basename(filePath),
-        // 文件大小
-        fileSize: (await stat(filePath)).size / (1024 * 1024),
-        // 元信息
-        common,
-        // 歌词
-        lyric:
-          metaDataLyricsArrayToLrc(common?.lyrics?.[0]?.syncText || []) ||
-          common?.lyrics?.[0]?.text ||
-          "",
-        // 音质信息
-        format,
-        // md5
-        md5: await getFileMD5(filePath),
-        replayGain: {
-          trackGain: common.replaygain_track_gain?.ratio,
-          trackPeak: common.replaygain_track_peak?.ratio,
-          albumGain: common.replaygain_album_gain?.ratio,
-          albumPeak: common.replaygain_album_peak?.ratio,
-        },
-      };
-    } catch (error) {
-      ipcLog.error("❌ Error fetching music metadata:", error);
-      throw error;
-    }
+    return musicFileService.getMusicMetadata(path);
   });
 
   // 修改音乐元信息
   ipcMain.handle("set-music-metadata", async (_, path: string, metadata: any) => {
-    try {
-      const {
-        name,
-        artist,
-        album,
-        alia,
-        lyric,
-        cover,
-        albumArtist,
-        genre,
-        year,
-        trackNumber,
-        discNumber,
-      } = metadata;
-      // 规范化路径
-      const songPath = resolve(path);
-      const coverPath = cover ? resolve(cover) : undefined;
-
-      const meta: SongMetadata = {
-        title: name || "未知曲目",
-        artist: artist || "未知艺术家",
-        album: album || "未知专辑",
-        lyric: lyric || "",
-        description: alia || "",
-        albumArtist: albumArtist,
-        genre: genre,
-        year: year,
-        trackNumber: trackNumber,
-        discNumber: discNumber,
-      };
-
-      if (!tools) {
-        throw new Error("Native tools not loaded");
-      }
-
-      await tools.writeMusicMetadata(songPath, meta, coverPath);
-      return true;
-    } catch (error) {
-      ipcLog.error("❌ Error setting music metadata:", error);
-      throw error;
-    }
+    return musicFileService.setMusicMetadata(path, metadata);
   });
 
   // 获取音乐歌词
-  ipcMain.handle(
-    "get-music-lyric",
-    async (
-      _,
-      musicPath: string, // 参数名改为 musicPath 以示区分
-    ): Promise<{
-      lyric: string;
-      format: "lrc" | "ttml" | "yrc";
-      external?: { lyric: string; format: "lrc" | "ttml" | "yrc" };
-      embedded?: { lyric: string; format: "lrc" };
-    }> => {
-      try {
-        // 获取文件基本信息
-        const absPath = resolve(musicPath);
-        const dir = dirname(absPath);
-        const ext = extname(absPath);
-        const baseName = basename(absPath, ext);
-        // 读取目录下所有文件
-        let files: string[] = [];
-        try {
-          files = await readdir(dir);
-        } catch (error) {
-          ipcLog.error("❌ Failed to read directory:", dir);
-          throw error;
-        }
-        // 外部歌词
-        let external: { lyric: string; format: "lrc" | "ttml" | "yrc" } | undefined;
-        // 内置歌词
-        let embedded: { lyric: string; format: "lrc" } | undefined;
-        // 查找外部歌词文件
-        for (const format of ["ttml", "yrc", "lrc"] as const) {
-          // 构造期望目标文件名
-          const targetNameLower = `${baseName}.${format}`.toLowerCase();
-          // 在文件列表中查找是否存在匹配项（忽略大小写）
-          const matchedFileName = files.find((file) => file.toLowerCase() === targetNameLower);
-          if (matchedFileName) {
-            try {
-              const lyricPath = join(dir, matchedFileName);
-              const lyric = await readFile(lyricPath, "utf-8");
-              // 若不为空
-              if (lyric && lyric.trim() !== "") {
-                ipcLog.info(`✅ Local lyric found (${format}): ${lyricPath}`);
-                external = { lyric, format };
-                break; // 找到最高优先级的外部歌词后停止
-              }
-            } catch {
-              // 读取失败则尝试下一种格式
-            }
-          }
-        }
-        // 读取内置元数据 (ID3 Tags)
-        try {
-          const { common } = await parseFile(absPath);
-          const syncedLyric = common?.lyrics?.[0]?.syncText;
-          if (syncedLyric && syncedLyric.length > 0) {
-            embedded = {
-              lyric: metaDataLyricsArrayToLrc(syncedLyric),
-              format: "lrc",
-            };
-          } else if (common?.lyrics?.[0]?.text) {
-            embedded = {
-              lyric: common?.lyrics?.[0]?.text,
-              format: "lrc",
-            };
-          }
-        } catch (e) {
-          ipcLog.warn(`⚠️ Failed to parse metadata for lyrics: ${absPath}`, e);
-        }
-        // 返回结果
-        const main = external || embedded || { lyric: "", format: "lrc" as const };
-        return {
-          ...main,
-          external,
-          embedded,
-        };
-      } catch (error) {
-        ipcLog.error("❌ Error fetching music lyric:", error);
-        throw error;
-      }
-    },
-  );
+  ipcMain.handle("get-music-lyric", async (_, musicPath: string) => {
+    return musicFileService.getMusicLyric(musicPath);
+  });
 
   // 获取音乐封面
-  ipcMain.handle(
-    "get-music-cover",
-    async (_, path: string): Promise<{ data: Buffer; format: string } | null> => {
-      try {
-        const { common } = await parseFile(path);
-        // 获取封面数据
-        const picture = common.picture?.[0];
-        if (picture) {
-          return { data: Buffer.from(picture.data), format: picture.format };
-        } else {
-          const coverFilePath = path.replace(/\.[^.]+$/, ".jpg");
-          try {
-            await access(coverFilePath);
-            const coverData = await readFile(coverFilePath);
-            return { data: coverData, format: "image/jpeg" };
-          } catch {
-            return null;
-          }
-        }
-      } catch (error) {
-        console.error("❌ Error fetching music cover:", error);
-        throw error;
-      }
-    },
-  );
+  ipcMain.handle("get-music-cover", async (_, path: string) => {
+    return musicFileService.getMusicCover(path);
+  });
 
   // 读取本地歌词
-  ipcMain.handle(
-    "read-local-lyric",
-    async (_, lyricDirs: string[], id: number): Promise<{ lrc: string; ttml: string }> => {
-      const result = { lrc: "", ttml: "" };
-
-      try {
-        // 定义需要查找的模式
-        // 此处的 `{,*.}` 表示这里可以取 `` (empty) 也可以取 `*.`
-        // 将歌词文件命名为 `歌曲ID.后缀名` 或者 `任意前缀.歌曲ID.后缀名` 均可
-        const patterns = {
-          ttml: `**/{,*.}${id}.ttml`,
-          lrc: `**/{,*.}${id}.lrc`,
-        };
-
-        // 遍历每一个目录
-        for (const dir of lyricDirs) {
-          try {
-            // 查找 ttml
-            if (!result.ttml) {
-              const ttmlFiles = await FastGlob(patterns.ttml, globOpt(dir));
-              if (ttmlFiles.length > 0) {
-                const filePath = join(dir, ttmlFiles[0]);
-                await access(filePath);
-                result.ttml = await readFile(filePath, "utf-8");
-              }
-            }
-
-            // 查找 lrc
-            if (!result.lrc) {
-              const lrcFiles = await FastGlob(patterns.lrc, globOpt(dir));
-              if (lrcFiles.length > 0) {
-                const filePath = join(dir, lrcFiles[0]);
-                await access(filePath);
-                result.lrc = await readFile(filePath, "utf-8");
-              }
-            }
-
-            // 如果两种文件都找到了就提前结束搜索
-            if (result.ttml && result.lrc) break;
-          } catch {
-            // 某个路径异常，跳过
-          }
-        }
-      } catch {
-        /* 忽略错误 */
-      }
-
-      return result;
-    },
-  );
+  ipcMain.handle("read-local-lyric", async (_, lyricDirs: string[], id: number) => {
+    return musicFileService.readLocalLyric(lyricDirs, id);
+  });
 
   // 删除文件
   ipcMain.handle("delete-file", async (_, path: string) => {
-    try {
-      // 规范化路径
-      const resolvedPath = resolve(path);
-      // 检查文件是否存在
-      try {
-        await access(resolvedPath);
-      } catch {
-        throw new Error("❌ File not found");
-      }
-      // 删除文件
-      await unlink(resolvedPath);
-      return true;
-    } catch (error) {
-      ipcLog.error("❌ File delete error", error);
-      return false;
-    }
+    return fileService.deleteFile(path);
   });
 
   // 打开文件夹
   ipcMain.on("open-folder", async (_, path: string) => {
-    try {
-      // 规范化路径
-      const resolvedPath = resolve(path);
-      // 检查文件夹是否存在
-      try {
-        await access(resolvedPath);
-      } catch {
-        throw new Error("❌ Folder not found");
-      }
-      // 打开文件夹
-      shell.showItemInFolder(resolvedPath);
-    } catch (error) {
-      ipcLog.error("❌ Folder open error", error);
-      throw error;
-    }
+    await fileService.openFolder(path);
   });
 
   // 图片选择窗口
   ipcMain.handle("choose-image", async () => {
-    try {
-      const { filePaths } = await dialog.showOpenDialog({
-        properties: ["openFile"],
-        filters: [{ name: "Images", extensions: ["jpg", "jpeg", "png"] }],
-      });
-      if (!filePaths || filePaths.length === 0) return null;
-      return filePaths[0];
-    } catch (error) {
-      ipcLog.error("❌ Image choose error", error);
-      return null;
-    }
+    return fileService.chooseImage();
   });
 
   // 路径选择窗口
   ipcMain.handle("choose-path", async (_, title: string, multiSelect: boolean = false) => {
-    try {
-      const properties: ("openDirectory" | "createDirectory" | "multiSelections")[] = [
-        "openDirectory",
-        "createDirectory",
-      ];
-      if (multiSelect) {
-        properties.push("multiSelections");
-      }
-      const { filePaths } = await dialog.showOpenDialog({
-        title: title ?? "选择文件夹",
-        defaultPath: app.getPath("downloads"),
-        properties,
-        buttonLabel: "选择文件夹",
-      });
-      if (!filePaths || filePaths.length === 0) return null;
-      // 多选时返回数组，单选时返回第一个路径
-      return multiSelect ? filePaths : filePaths[0];
-    } catch (error) {
-      ipcLog.error("❌ Path choose error", error);
-      return null;
-    }
+    return fileService.choosePath(title, multiSelect);
   });
 
   // 下载文件
-  ipcMain.handle(
-    "download-file",
-    async (
-      event,
-      url: string,
-      options: {
-        fileName: string;
-        fileType: string;
-        path: string;
-        downloadMeta?: boolean;
-        downloadCover?: boolean;
-        downloadLyric?: boolean;
-        saveMetaFile?: boolean;
-        lyric?: string;
-        songData?: any;
-        skipIfExist?: boolean;
-        threadCount?: number;
-        referer?: string;
-        enableDownloadHttp2?: boolean;
-      } = {
-        fileName: "未知文件名",
-        fileType: "mp3",
-        path: app.getPath("downloads"),
-      },
-    ): Promise<{ status: "success" | "skipped" | "error" | "cancelled"; message?: string }> => {
-      try {
-        // 获取窗口
-        const win = BrowserWindow.fromWebContents(event.sender);
-        if (!win) return { status: "error", message: "Window not found" };
-        // 获取配置
-        const {
-          fileName,
-          fileType,
-          path,
-          lyric,
-          downloadMeta,
-          downloadCover,
-          downloadLyric,
-          saveMetaFile,
-          songData,
-          skipIfExist,
-          referer,
-          enableDownloadHttp2,
-        } = options;
-        // 规范化路径
-        const downloadPath = resolve(path);
-        // 检查文件夹是否存在，不存在则自动递归创建
-        try {
-          await access(downloadPath);
-        } catch {
-          await mkdir(downloadPath, { recursive: true });
-        }
+  ipcMain.handle("download-file", async (event, url: string, options: any) => {
+    const win = BrowserWindow.fromWebContents(event.sender);
+    if (!win) return { status: "error", message: "Window not found" };
 
-        const finalFilePath = fileType
-          ? join(downloadPath, `${fileName}.${fileType}`)
-          : join(downloadPath, fileName);
+    // Set default options if not provided
+    const downloadOptions = {
+      fileName: "未知文件名",
+      fileType: "mp3",
+      path: fileService.getDefaultDir("downloads"),
+      ...options,
+    };
 
-        // 检查文件是否存在
-        if (skipIfExist) {
-          try {
-            await access(finalFilePath);
-            return { status: "skipped", message: "文件已存在" };
-          } catch {
-            // 文件不存在，继续下载
-          }
-        }
-
-        // Prepare metadata
-        let metadata: SongMetadata | undefined | null = null;
-        if (downloadMeta && songData) {
-          const getArtistNames = (artists: any): string[] => {
-            if (Array.isArray(artists)) {
-              return artists
-                .map((ar: any) => (typeof ar === "string" ? ar : ar?.name || ""))
-                .filter((name) => name && name.trim().length > 0);
-            }
-            if (typeof artists === "string" && artists.trim().length > 0) {
-              return [artists];
-            }
-            return [];
-          };
-          const artistNames = getArtistNames(songData.artists);
-          const artist = artistNames.join(", ") || "未知艺术家";
-
-          const coverUrl =
-            downloadCover && (songData.coverSize?.l || songData.cover)
-              ? songData.coverSize?.l || songData.cover
-              : undefined;
-
-          metadata = {
-            title: songData.name || "未知曲目",
-            artist: artist,
-            album:
-              (typeof songData.album === "string" ? songData.album : songData.album?.name) ||
-              "未知专辑",
-            coverUrl: coverUrl,
-            lyric: downloadLyric && lyric ? lyric : undefined,
-            description: songData.alia || "",
-          };
-          // console.log("[Download] Resolved cover URL:", coverUrl);
-        }
-
-        const onProgress = (...args: any[]) => {
-          let progressData: any;
-
-          // Handle (err, value) or (value) signature
-          if (args.length > 1 && args[0] === null) {
-            progressData = args[1];
-          } else if (args.length > 0) {
-            progressData = args[0];
-          }
-
-          try {
-            if (!progressData) return;
-
-            // Handle both object (new) and JSON string (legacy/fallback)
-            if (typeof progressData === "string") {
-              try {
-                progressData = JSON.parse(progressData);
-              } catch (e) {
-                console.error("Failed to parse progress json", e);
-                return;
-              }
-            }
-
-            if (!progressData || typeof progressData !== "object") return;
-
-            // Map snake_case (Rust) to camelCase (JS)
-            // Rust struct: { percent, transferred_bytes, total_bytes }
-            const percent = progressData.percent;
-            const transferredBytes =
-              progressData.transferredBytes ?? progressData.transferred_bytes ?? 0;
-            const totalBytes = progressData.totalBytes ?? progressData.total_bytes ?? 0;
-
-            win.webContents.send("download-progress", {
-              id: songData?.id,
-              percent: percent,
-              transferredBytes: transferredBytes,
-              totalBytes: totalBytes,
-            });
-          } catch (e) {
-            console.error("Error processing progress callback", e, "Args:", args);
-          }
-        };
-
-        if (!tools) {
-          throw new Error("Native tools not loaded");
-        }
-
-        const store = useStore();
-        // Use threadCount from options if available, otherwise fall back to store
-        const threadCount =
-          (options.threadCount as number) || (store.get("downloadThreadCount") as number) || 8;
-
-        const enableHttp2 =
-          enableDownloadHttp2 !== undefined
-            ? enableDownloadHttp2
-            : (store.get("enableDownloadHttp2", true) as boolean);
-
-        // Upgrade HTTP to HTTPS if HTTP2 is enabled (HTTP2 usually requires HTTPS)
-        let finalUrl = url;
-        if (enableHttp2 && finalUrl.startsWith("http://")) {
-          finalUrl = finalUrl.replace(/^http:\/\//, "https://");
-          ipcLog.info(`🔒 Upgraded download URL to HTTPS for HTTP/2 support: ${finalUrl}`);
-        }
-
-        const task = new tools.DownloadTask();
-        const downloadId = songData?.id || 0;
-        activeDownloads.set(downloadId, task);
-
-        try {
-          await task.download(
-            finalUrl,
-            finalFilePath,
-            metadata,
-            threadCount,
-            referer,
-            onProgress,
-            enableHttp2,
-          );
-        } finally {
-          activeDownloads.delete(downloadId);
-        }
-
-        // 创建同名歌词文件
-        if (lyric && saveMetaFile && downloadLyric) {
-          const lrcPath = join(downloadPath, `${fileName}.lrc`);
-          await writeFile(lrcPath, lyric, "utf-8");
-        }
-
-        return { status: "success" };
-      } catch (error: any) {
-        ipcLog.error("❌ Error downloading file:", error);
-        if ((error.message && error.message.includes("cancelled")) || error.code === "Cancelled") {
-          return { status: "cancelled", message: "下载已取消" };
-        }
-        return {
-          status: "error",
-          message: error instanceof Error ? error.message : "Unknown error",
-        };
-      }
-    },
-  );
+    return downloadService.downloadFile(url, downloadOptions, win);
+  });
 
   // 取消下载
   ipcMain.handle("cancel-download", async (_, songId: number) => {
-    const task = activeDownloads.get(songId);
-    if (task) {
-      task.cancel();
-      return true;
-    }
-    return false;
+    return downloadService.cancelDownload(songId);
   });
 
-  // 检查是否是相同的路径（规范化后比较）
+  // 检查是否是相同的路径
   ipcMain.handle("check-if-same-path", (_, localFilesPath: string[], selectedDir: string) => {
-    const resolvedSelectedDir = resolve(selectedDir);
-    const allPaths = localFilesPath.map((p) => resolve(p));
-    return allPaths.some((existingPath) => existingPath === resolvedSelectedDir);
+    return fileService.checkIfSamePath(localFilesPath, selectedDir);
   });
 
   // 检查是否是子文件夹
   ipcMain.handle("check-if-subfolder", (_, localFilesPath: string[], selectedDir: string) => {
-    const resolvedSelectedDir = resolve(selectedDir);
-    const allPaths = localFilesPath.map((p) => resolve(p));
-    return allPaths.some((existingPath) => {
-      const relativePath = relative(existingPath, resolvedSelectedDir);
-      return relativePath && !relativePath.startsWith("..") && !isAbsolute(relativePath);
-    });
+    return fileService.checkIfSubfolder(localFilesPath, selectedDir);
   });
 
-  // 保存文件内容 (用于保存文本文件等)
-  ipcMain.handle(
-    "save-file-content",
-    async (
-      _,
-      options: { path: string; fileName: string; content: string; encoding?: string },
-    ): Promise<{ success: boolean; message?: string }> => {
-      try {
-        const { path, fileName, content, encoding = "utf-8" } = options;
-        // 规范化路径
-        const dirPath = resolve(path);
-        // 检查文件夹是否存在，不存在则自动递归创建
-        try {
-          await access(dirPath);
-        } catch {
-          await mkdir(dirPath, { recursive: true });
-        }
-        const filePath = join(dirPath, fileName);
-
-        if (encoding !== "utf-8") {
-          try {
-            // 使用动态导入，避免启动时加载问题
-            const { encode } = await import("iconv-lite");
-            // iconv-lite support 'utf16' as alias for 'utf-16' etc.
-            const buffer = encode(content, encoding);
-            await writeFile(filePath, buffer);
-          } catch (e) {
-            ipcLog.error(`❌ ${encoding} encoding failed:`, e);
-            // Fallback to UTF-8 on error
-            await writeFile(filePath, content, "utf-8");
-          }
-        } else {
-          await writeFile(filePath, content, "utf-8");
-        }
-
-        return { success: true };
-      } catch (error) {
-        ipcLog.error("❌ Error saving file content:", error);
-        return {
-          success: false,
-          message: error instanceof Error ? error.message : "Unknown error",
-        };
-      }
-    },
-  );
+  // 保存文件内容
+  ipcMain.handle("save-file-content", async (_, options: any) => {
+    return fileService.saveFileContent(options);
+  });
 };
 
 function* chunkArray<T>(array: T[], size: number): Generator<T[], void, unknown> {
